% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppellsF1.R
\name{AppellsF1}
\alias{AppellsF1}
\title{The Appell's first hypergeometric function of two variables x and y.}
\usage{
AppellsF1(a, b1, b2, c, x, y)
}
\arguments{
\item{a}{A numeric value.}

\item{b1}{A numeric value.}

\item{b2}{A numeric value.}

\item{c}{A numeric value.}

\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The \code{AppellsF1} gives the numeric value of the Appell's first hypergeometric function.
}
\description{
This function calculates the numeric value of the Appell's first hypergeometric function.
}
\examples{
AppellsF1(0.5, 0.5, 0, 1, 0.96, 1.2)

}
