% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDecisionProbBinary.R
\name{BayesDecisionProbBinary}
\alias{BayesDecisionProbBinary}
\title{Calculate the Go, NoGo and Gray probabilities for a clinical trial when outcome is binary
under the Bayesian framework using two metrics:
(i)  posterior probability for the treatment effect to be greater than a threshold, and
(ii) posterior predictive probability of phase III study success.}
\usage{
BayesDecisionProbBinary(
  prob,
  design,
  theta.TV,
  theta.MAV,
  theta.NULL,
  gamma1,
  gamma2,
  pi1,
  pi2,
  n1,
  n2,
  a1,
  a2,
  b1,
  b2,
  z,
  m1,
  m2,
  ne1,
  ne2,
  ye1,
  ye2,
  ae1,
  ae2
)
}
\arguments{
\item{prob}{A type of probability (i.e., \code{prob='posterior'} or \code{prob='predictive'}).}

\item{design}{A type of design (i.e., \code{design='controlled'}, \code{design='uncontrolled'} or \code{design='external'}).}

\item{theta.TV}{A numeric pre-specified threshold value for calculating Go probability when \code{prob='posterior'}.}

\item{theta.MAV}{A numeric pre-specified threshold value for calculating NoGo probability when \code{prob='posterior'}.}

\item{theta.NULL}{A numeric pre-specified threshold value for calculating Go/NoGo probabilities when \code{prob='predictive'}.}

\item{gamma1}{A numeric value of a minimum probability to declare success.}

\item{gamma2}{A numeric value of a futility threshold.}

\item{pi1}{A numeric value or a vector representing true response probability(s) for group 1.}

\item{pi2}{A numeric value or a vector representing true response probability(s) for group 2.}

\item{n1}{A number of patients in group 1 for a proof-of-concept (PoC) trial.}

\item{n2}{A number of patients in group 2 for the PoC trial.}

\item{a1}{The first non-negative numeric value of the prior distribution for group 1.}

\item{a2}{The first non-negative numeric value of the prior distribution for group 2.}

\item{b1}{The second non-negative numeric value of the prior distribution for group 1.}

\item{b2}{The second non-negative numeric value of the prior distribution for group 2.}

\item{z}{A hypothetical observed number of responders in group 2 for an uncontrolled design.}

\item{m1}{A number of patients in group 1 for the future trial data.}

\item{m2}{A number of patients in group 2 for the future trial data.}

\item{ne1}{A number of patients in group 1 for the external data.}

\item{ne2}{A number of patients in group 2 for the external data.}

\item{ye1}{An observed number of responders in group 1 for the external data.}

\item{ye2}{An observed number of responders in group 2 for the external data.}

\item{ae1}{A scale parameter (power parameter) for group 1.}

\item{ae2}{A scale parameter (power parameter) for group 2.}
}
\value{
The \code{BayesDecisionProbBinary} gives the dataframe including the
true response probabilities for both groups, and the Go, NoGo and Gray probabilities.
}
\description{
The function can obtain:
\itemize{
\item Go probability
\item NoGo probability
\item Gray probability
}
The function can be used for controlled design, uncontrolled design and design using an external (historical) data.
}
\examples{
BayesDecisionProbBinary(
  prob = 'posterior', design = 'controlled', theta.TV = 0.4, theta.MAV = 0.2, theta.NULL = NULL, gamma1 = 0.5, gamma2 = 0.2, 
  pi1 = c(0.2, 0.4, 0.6, 0.8), pi2 = rep(0.2, 4), n1 = 12, n2 = 12, a1 = 0.5, a2 = 0.5, b1 = 0.5, b2 = 0.5, z = NULL,
  m1 = NULL, m2 = NULL, ne1 = NULL, ne2 = NULL, ye1 = NULL, ye2 = NULL, ae1 = NULL, ae2 = NULL
)
BayesDecisionProbBinary(
  prob = 'predictive', design = 'controlled', theta.TV = NULL, theta.MAV = NULL, theta.NULL = 0, gamma1 = 0.9, gamma2 = 0.3, 
  pi1 = c(0.2, 0.4, 0.6, 0.8), pi2 = rep(0.2, 4), n1 = 12, n2 = 12, a1 = 0.5, a2 = 0.5, b1 = 0.5, b2 = 0.5, z = NULL,
  m1 = 30, m2 = 30, ne1 = NULL, ne2 = NULL, ye1 = NULL, ye2 = NULL, ae1 = NULL, ae2 = NULL
)

}
