% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPostPredContinuous.R
\name{BayesPostPredContinuous}
\alias{BayesPostPredContinuous}
\title{Calculate bayesian posterior probability or bayesian posterior predictive probability
for a clinical trial when outcome is continuous.}
\usage{
BayesPostPredContinuous(
  prob = "posterior",
  design = "controlled",
  prior = "vague",
  CalcMethod = "NI",
  theta0,
  nMC = 10000,
  n1,
  n2,
  m1,
  m2,
  kappa01,
  kappa02,
  nu01,
  nu02,
  mu01,
  mu02,
  sigma01,
  sigma02,
  bar.y1,
  bar.y2,
  s1,
  s2,
  r
)
}
\arguments{
\item{prob}{A type of probability (i.e., \code{prob='posterior'} or \code{prob='predictive'}).}

\item{design}{A type of trial design (i.e., \code{design='controlled'} or \code{prob='uncontrolled'}).}

\item{prior}{A prior distribution (i.e., \code{prior='N-Inv-Chisq'} or \code{prior='vague'}).}

\item{CalcMethod}{An option to select a calculation method (\code{CalcMethod=NI} (numerical integration method), \code{CalcMethod=MC} (Monte Carlo method), or \code{CalcMethod=WS} (WS approximation)).}

\item{theta0}{Numeric pre-specified threshold value.}

\item{nMC}{A number of iterations for Monte Carlo simulation (a numeric value is set only if \code{CalcMethod=MC}).}

\item{n1}{A number of patients in group 1 for a proof-of-concept (PoC) trial.}

\item{n2}{A number of patients in group 2 for the PoC trial.}

\item{m1}{A number of patients in group 1 for the future trial data.}

\item{m2}{A number of patients in group 2 for the future trial data.}

\item{kappa01}{A true value related to the mean for conjugate prior of Normal-Inverse-X^2 in group 1.}

\item{kappa02}{A true value related to the mean for conjugate prior of Normal-Inverse-X^2 in group 2.}

\item{nu01}{A true value related to the variance for conjugate prior of Normal-Inverse-X^2 in group 1.}

\item{nu02}{A true value related to the variance for conjugate prior of Normal-Inverse-X^2 in group 2.}

\item{mu01}{A true mean value of outcomes in group 1 for the PoC trial.}

\item{mu02}{A true mean value of outcomes in group 2 for the PoC trial.}

\item{sigma01}{A true standard deviation value of outcomes in group 1 for the PoC trial.}

\item{sigma02}{A true standard deviation value of outcomes in group 2 for the PoC trial.}

\item{bar.y1}{A sample mean of group 1.}

\item{bar.y2}{A sample mean of group 2.}

\item{s1}{A sample standard deviation of group 1.}

\item{s2}{A sample standard deviation of group 2.}

\item{r}{A parameter value associated with the distribution of mean for group 2 When \code{prob='uncontrolled'}.}
}
\value{
The \code{BayesPostPredContinuous} gives the numeric value(s) of the
bayesian posterior probability or bayesian posterior predictive probability.
}
\description{
The function can obtain:
\itemize{
\item Bayesian posterior probability
\item Bayesian posterior predictive probability
}
Prior distribution of mean and variance of outcomes for each treatment group (k=1,2) is (1) Normal-Inverse-X^2 or (2) Vague.
Posterior distribution or posterior predictive distribution of outcome for each treatment group is following t-distribution.
}
\examples{
sapply(c(2, 0), function(theta0) {
  BayesPostPredContinuous(
    prob = 'posterior', design = 'controlled', prior = 'N-Inv-Chisq', CalcMethod = 'NI',
    nMC = NULL, theta0 = theta0, n1 = 12, n2 = 12, m1 = NULL, m2 = NULL,
    kappa01 = 5, kappa02 = 5, nu01 = 5, nu02 = 5, mu01 = 5, mu02 = 5,
    sigma01 = sqrt(5), sigma02 = sqrt(5), bar.y1 = 2, bar.y2 = 0, s1 = 1, s2 = 1, r = NULL
  )
})
sapply(c(2, 0), function(theta0) {
  BayesPostPredContinuous(
    prob = 'posterior', design = 'controlled', prior = 'N-Inv-Chisq', CalcMethod = 'MC',
    nMC = 1e+6, theta0 = theta0, n1 = 12, n2 = 12, m1 = NULL, m2 = NULL,
    kappa01 = 5, kappa02 = 5, nu01 = 5, nu02 = 5, mu01 = 5, mu02 = 5,
    sigma01 = sqrt(5), sigma02 = sqrt(5), bar.y1 = 2, bar.y2 = 0, s1 = 1, s2 = 1, r = NULL
  )
})
sapply(c(2, 0), function(theta0) {
  BayesPostPredContinuous(
    prob = 'posterior', design = 'controlled', prior = 'N-Inv-Chisq', CalcMethod = 'WS',
    nMC = NULL, theta0 = theta0, n1 = 12, n2 = 12, m1 = NULL, m2 = NULL,
    kappa01 = 5, kappa02 = 5, nu01 = 5, nu02 = 5, mu01 = 5, mu02 = 5,
    sigma01 = sqrt(5), sigma02 = sqrt(5), bar.y1 = 2, bar.y2 = 0, s1 = 1, s2 = 1, r = NULL
  )
})
BayesPostPredContinuous(
  prob = 'predictive', design = 'uncontrolled', prior = 'vague', CalcMethod = 'NI',
  nMC = NULL, theta0 = 0.5, n1 = 12, n2 = NULL, m1 = 120, m2 = 120,
  kappa01 = NULL, kappa02 = NULL, nu01 = NULL, nu02 = NULL, mu01 = NULL, mu02 = 0,
  sigma01 = NULL, sigma02 = NULL, bar.y1 = 2, bar.y2 = NULL, s1 = 1, s2 = NULL, r = 12
)
BayesPostPredContinuous(
  prob = 'predictive', design = 'uncontrolled', prior = 'vague', CalcMethod = 'MC',
  nMC = 1e+6, theta0 = 0.5, n1 = 12, n2 = NULL, m1 = 120, m2 = 120,
  kappa01 = NULL, kappa02 = NULL, nu01 = NULL, nu02 = NULL, mu01 = NULL, mu02 = 0,
  sigma01 = NULL, sigma02 = NULL, bar.y1 = 2, bar.y2 = NULL, s1 = 1, s2 = NULL, r = 12
)
BayesPostPredContinuous(
  prob = 'predictive', design = 'uncontrolled', prior = 'vague', CalcMethod = 'WS',
  nMC = NULL, theta0 = 0.5, n1 = 12, n2 = NULL, m1 = 120, m2 = 120,
  kappa01 = NULL, kappa02 = NULL, nu01 = NULL, nu02 = NULL, mu01 = NULL, mu02 = 0,
  sigma01 = NULL, sigma02 = NULL, bar.y1 = 2, bar.y2 = NULL, s1 = 1, s2 = NULL, r = 12
)

}
