% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff2beta.R
\name{ddiff2beta}
\alias{ddiff2beta}
\title{Density of difference between two beta variables}
\usage{
ddiff2beta(pi, alpha1, alpha2, beta1, beta2)
}
\arguments{
\item{pi}{A numeric value representing the difference of the two beta variables.}

\item{alpha1}{The first non-negative numeric value of beta function for group 1.}

\item{alpha2}{The first non-negative numeric value of beta function for group 2.}

\item{beta1}{The second non-negative numeric value of beta function for group 1.}

\item{beta2}{The second non-negative numeric value of beta function for group 2.}
}
\value{
The \code{ddiff2beta} gives the numeric value of the density.
}
\description{
This function calculates density of difference between two beta variables
using Appell’s first hypergeometric function of two variables.
}
\examples{
ddiff2beta(0.2, 0.5, 0.5, 0.5, 0.5)

}
