% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPostPredBinary.R
\name{BayesPostPredBinary}
\alias{BayesPostPredBinary}
\title{Calculate bayesian posterior probability or bayesian posterior predictive probability
for a clinical trial when outcome is binary.}
\usage{
BayesPostPredBinary(
  prob,
  external,
  theta0,
  n1,
  n2,
  y1,
  y2,
  a1,
  a2,
  b1,
  b2,
  m1,
  m2,
  ne1,
  ne2,
  ye1,
  ye2,
  ae1,
  ae2
)
}
\arguments{
\item{prob}{A type of probability (i.e., \code{prob='posterior'} or \code{prob='predictive'}).}

\item{external}{When external data is used, \code{external=TRUE}.}

\item{theta0}{A numeric pre-specified threshold value.}

\item{n1}{A number of patients in group 1 for a proof-of-concept (PoC) trial.}

\item{n2}{A number of patients in group 2 for the PoC trial.}

\item{y1}{An observed number of responders in group 1 for the PoC trial.}

\item{y2}{An observed number of responders in group 2 for the PoC trial.}

\item{a1}{The first non-negative numeric value of the prior distribution for group 1.}

\item{a2}{The first non-negative numeric value of the prior distribution for group 2.}

\item{b1}{The second non-negative numeric value of the prior distribution for group 1.}

\item{b2}{The second non-negative numeric value of the prior distribution for group 2.}

\item{m1}{A number of patients in group 1 for the future trial data.}

\item{m2}{A number of patients in group 2 for the future trial data.}

\item{ne1}{A number of patients in group 1 for the external data.}

\item{ne2}{A number of patients in group 2 for the external data.}

\item{ye1}{An observed number of responders in group 1 for the external data.}

\item{ye2}{An observed number of responders in group 2 for the external data.}

\item{ae1}{A scale parameter (power parameter) for group 1.}

\item{ae2}{A scale parameter (power parameter) for group 2.}
}
\value{
The \code{BayesPostPredBinary} gives the numeric value of the
bayesian posterior probability or bayesian posterior predictive probability.
}
\description{
The function can obtain:
\itemize{
\item Bayesian posterior probability
\item Bayesian posterior predictive probability
}
Prior distribution of proportion of responders (pij) for each treatment group (j=1,2) is
following beta distribution.
For posterior probability, posterior distribution of pij is following beta distribution.
For posterior predictive probability, predictive distribution of future trial data is
following beta-binomial distribution.
The function can account for external (historical) data.
}
\examples{
BayesPostPredBinary(
  prob = 'posterior', external = TRUE, theta0 = 0.15,
  n1 = 12, n2 = 15, y1 = 7, y2 = 9, a1 = 0.5, a2 = 0.5, b1 = 0.5, b2 = 0.5,
  m1 = NULL, m2 = NULL, ne1 = 12, ne2 = 12, ye1 = 6, ye2 = 6, ae1 = 0.5, ae2 = 0.5
)
BayesPostPredBinary(
  prob = 'predictive', external = TRUE, theta0 = 0.5,
  n1 = 12, n2 = 15, y1 = 7, y2 = 7, a1 = 0.5, a2 = 0.5, b1 = 0.5, b2 = 0.5,
  m1 = 12, m2 = 12, ne1 = 12, ne2 = 12, ye1 = 6, ye2 = 6, ae1 = 0.5, ae2 = 0.5
)

}
