% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBetaBinomdiff.R
\name{pBetaBinomdiff}
\alias{pBetaBinomdiff}
\title{Cumulative distribution function (CDF) of the difference between two beta-binomial variables}
\usage{
pBetaBinomdiff(q, m1, m2, alpha1, alpha2, beta1, beta2)
}
\arguments{
\item{q}{Quantile value.}

\item{m1}{A number of patients in group 1 for the future trial data.}

\item{m2}{A number of patients in group 2 for the future trial data.}

\item{alpha1}{The first non-negative numeric value of beta function for group 1.}

\item{alpha2}{The first non-negative numeric value of beta function for group 2.}

\item{beta1}{The second non-negative numeric value of beta function for group 1.}

\item{beta2}{The second non-negative numeric value of beta function for group 2.}
}
\value{
The \code{pBetaBinomdiff} gives the numeric value of the density.
}
\description{
This function calculates CDF(s) of difference between two beta-binomial variables.
}
\examples{
pBetaBinomdiff(0.2, 12, 12, 0.5, 0.5, 0.5, 0.5)

}
