% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBetadiff.R
\name{pBetadiff}
\alias{pBetadiff}
\title{Cumulative distribution function (CDF) of the difference between two beta variables}
\usage{
pBetadiff(q, alpha1, alpha2, beta1, beta2)
}
\arguments{
\item{q}{Quantile value.}

\item{alpha1}{The first non-negative numeric value of beta function for group 1.}

\item{alpha2}{The first non-negative numeric value of beta function for group 2.}

\item{beta1}{The second non-negative numeric value of beta function for group 1.}

\item{beta2}{The second non-negative numeric value of beta function for group 2.}
}
\value{
The \code{pBetadiff} gives the numeric value of the density.
}
\description{
This function calculates CDF(s) of difference between two beta variables
using Appell’s first hypergeometric function of two variables.
}
\examples{
pBetadiff(0.2, 0.5, 0.5, 0.5, 0.5)

}
